создание контроллера для поиска
